# Refactor enums comparison operators for performance improvements