… src/runtime/Properties/AssemblyInfo.cs
… src/runtime/Util/OpsHelper.cs